Extract duplicated helpers into a shared utils module
deviation_exceeds, update_reputation, apply_reputation_decay,
calculate_usd_price, vec_contains_address, and append_history (plus the
reputation constants) were copy-pasted between contract.rs and proxy.rs
with drift — the two calculate_usd_price implementations used different
math. Move them into a single crate::utils module used by both contracts,
and unify USD conversion to one documented formula with checked
arithmetic (returns None on overflow instead of panicking).
