+++ contracts/price-oracle/src/proxy.rs
@@ -3,6 +3,10 @@ use soroban_sdk::{contract, contractimpl, Address, BytesN, Env, String, Vec};
 use crate::errors::OracleError;
 use crate::storage;
 use crate::types::{AssetPrice, MultiSigConfig, PriceDataPoint, SourceReputation};
+use crate::utils::{append_history, apply_reputation_decay, calculate_usd_price, deviation_exceeds, update_reputation, vec_contains_address};
+
+// Issue #375 — minimum delay between a queued WASM upgrade and its execution.
+const UPGRADE_TIMELOCK_SECS: u64 = 172_800; // 48 hours
 
 // Issue #68: Proxy contract with upgradeability via WASM hash replacement.
 //
@@ -342,9 +346,7 @@ impl ProxyContract {
 
         storage::set_latest_price(&env, &asset, &data_point);
 
-        let mut history = storage::get_price_history(&env, &asset);
-        history.push_back(data_point.clone());
-        storage::set_price_history(&env, &asset, &history);
+        append_history(&env, &asset, data_point.clone());
 
         env.events()
             .publish(("price_submitted", asset, source), (price, timestamp));
@@ -421,97 +423,3 @@ impl ProxyContract {
         Ok(())
     }
 }
-
-// -------------------------------------------------------------------------
-// Shared helpers (duplicated from contract.rs; cannot call across contracts
-// within the same crate without cross-contract invocation overhead)
-// -------------------------------------------------------------------------
-
-const REPUTATION_ACCURACY_THRESHOLD_BPS: u128 = 2000;
-const REPUTATION_DECAY_PERIOD_SECS: u64 = 604_800;
-// Issue #375 — minimum delay between a queued WASM upgrade and its execution.
-const UPGRADE_TIMELOCK_SECS: u64 = 172_800; // 48 hours
-
-fn vec_contains_address(vec: &Vec<Address>, target: &Address) -> bool {
-    for i in 0..vec.len() {
-        if let Some(addr) = vec.get(i) {
-            if &addr == target {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-fn deviation_exceeds(new_price: i128, prev_price: i128, threshold_bps: u32) -> bool {
-    if prev_price == 0 {
-        return false;
-    }
-    let prev_abs = prev_price.unsigned_abs();
-    let diff: u128 = if (new_price >= 0) == (prev_price >= 0) {
-        let new_abs = new_price.unsigned_abs();
-        if new_abs >= prev_abs { new_abs - prev_abs } else { prev_abs - new_abs }
-    } else {
-        new_price.unsigned_abs().saturating_add(prev_abs)
-    };
-    diff.saturating_mul(10_000) / prev_abs > threshold_bps as u128
-}
-
-fn update_reputation(env: &Env, source: &Address, new_price: i128, asset: &String, timestamp: u64) {
-    let is_accurate = match storage::get_latest_price(env, asset) {
-        None => true,
-        Some(prev) => !deviation_exceeds(new_price, prev.price, REPUTATION_ACCURACY_THRESHOLD_BPS as u32),
-    };
-
-    let mut rep = storage::get_source_reputation(env, source).unwrap_or(SourceReputation {
-        score: 10_000,
-        total_submissions: 0,
-        accurate_submissions: 0,
-        last_updated: timestamp,
-    });
-
-    rep.total_submissions = rep.total_submissions.saturating_add(1);
-    if is_accurate {
-        rep.accurate_submissions = rep.accurate_submissions.saturating_add(1);
-    }
-    rep.score = if rep.total_submissions == 0 {
-        10_000
-    } else {
-        (rep.accurate_submissions as u32).saturating_mul(10_000) / rep.total_submissions
-    };
-    rep.last_updated = timestamp;
-    storage::set_source_reputation(env, source, &rep);
-}
-
-fn apply_reputation_decay(env: &Env, mut rep: SourceReputation) -> SourceReputation {
-    let now = env.ledger().timestamp();
-    let elapsed = now.saturating_sub(rep.last_updated);
-    if elapsed < REPUTATION_DECAY_PERIOD_SECS {
-        return rep;
-    }
-    let periods = (elapsed / REPUTATION_DECAY_PERIOD_SECS).min(40) as u32;
-    for _ in 0..periods {
-        rep.score = rep.score.saturating_mul(95) / 100;
-    }
-    rep
-}
-
-fn calculate_usd_price(env: &Env, asset: &String, price: i128, decimals: u32) -> Option<i128> {
-    let xlm = String::from_str(env, "XLM");
-    if asset == &xlm {
-        return Some(price);
-    }
-    let usdc = String::from_str(env, "USDC");
-    if let Some(usdc_anchor) = storage::get_latest_price(env, &usdc) {
-        if asset == &usdc {
-            return Some(10i128.pow(decimals));
-        }
-        if let Some(xlm_price) = storage::get_latest_price(env, &xlm) {
-            let base_asset_price = (price * xlm_price.price * 10i128.pow(decimals))
-                / (10i128.pow(decimals) * 10i128.pow(usdc_anchor.decimals));
-            return Some(base_asset_price);
-        }
-    }
-    None
-}
-
